+++ src/app/util/helper-functions.ts
@@ -11,3 +11,11 @@ export function dateAsUTCDate(date: Date) {
     d.setUTCMilliseconds(date.getMilliseconds());
     return d;
 }
+
+export function formatBytes(bytes, decimals = 2) {
+    if (bytes === 0) return '0 Bytes';
+    const dm = decimals < 0 ? 0 : decimals;
+    const sizes = ['Bytes', 'KB', 'MB', 'GB', 'TB', 'PB', 'EB', 'ZB', 'YB'];
+    const i = Math.floor(Math.log(bytes) / Math.log(1024));
+    return parseFloat((bytes / Math.pow(1024, i)).toFixed(dm)) + ' ' + sizes[i];
+}
